Lime3DS: Migrate to Tracy profiler from Microprofile
Original Description:
The current Microprofile used by Citra hasn't been updated for ages and is based on an abandoned fork of Microprofile. This PR removes it in favour of a new profiler, Tracy. This change has some implications to how Citra will be profiled in the future. As far as integration goes, the documentation recommends using tracy as a submodule, however given that repository contains many others things like the profile applications that we don't need, I've opted to manually include the necessary source files. They are self-contained in the public folder so updating them should be as easy as copying the relevant folder from the Tracy repo to Citra externals. An option to completely disable profiling has also been added.

Cons:

To use Tracy a separate application is required, while Microprofile had an integrated UI directly in Citra. This might make things harder in cases where developers require a user to profile the application, though I don't believe that is something often.

Another concern is performance compared to Microprofile, as Tracy is also a sampling profiler. In this pull request I enable on demand profiling so the cost shouldn't be noticeable when not profiling but this should be tested.

Pros:

Being an external program, Tracy has many more features compared to Microprofile. It has the normal frame time view like Microprofile but can give per-function information in most cases, while Microprofile only gives timing information for the defined scopes. In the below screenshot for example, I profiled the prototype Teak JIT compiler and Tracy can show which function exactly consumes the most time as well as any calls made.